pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-session>=0.8.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
]
